requirements.txt
aenum==3.1.12
attrs==22.1.0
beautifulsoup4==4.12.0
paho_mqtt==1.6.1
poe-api==0.4.3
python-dotenv==1.0.0
Requests==2.31.0